Rename files
The router timeouts aren't ordinary/emergency, not really, so naming
them like that is misleading
